+++ src/Internal/LazyResult.php
@@ -61,33 +61,33 @@ final class LazyResult extends BaseStreamPipe implements Result
     /**
      * @inheritDoc
      */
-    public function toString(): string
+    public function toString(string $separator = ','): string
     {
-        return $this->execute()->toString();
+        return $this->execute()->toString($separator);
     }
     
     /**
      * @inheritDoc
      */
-    public function toJson(): string
+    public function toJson(int $flags = 0, bool $preserveKeys = false): string
     {
-        return $this->execute()->toJson();
+        return $this->execute()->toJson($flags, $preserveKeys);
     }
     
     /**
      * @inheritDoc
      */
-    public function toJsonAssoc(): string
+    public function toJsonAssoc(int $flags = 0): string
     {
-        return $this->execute()->toJsonAssoc();
+        return $this->execute()->toJsonAssoc($flags);
     }
     
     /**
      * @inheritDoc
      */
-    public function toArray(): array
+    public function toArray(bool $preserveKeys = false): array
     {
-        return $this->execute()->toArray();
+        return $this->execute()->toArray($preserveKeys);
     }
     
     /**
@@ -112,14 +112,6 @@ final class LazyResult extends BaseStreamPipe implements Result
     public function call($consumer): void
     {
         $this->execute()->call($consumer);
-    }
-    
-    /**
-     * @inheritDoc
-     */
-    public function __toString(): string
-    {
-        return $this->execute()->toString();
     }
     
     private function execute(): Result